Renamed _functional/static_magic.h_ -> _functional/always_default.h_
